requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
distlib==0.3.1
distro==1.5.0
Django==3.1.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
environment==1.0.0
filelock==3.0.12
Jinja2==3.0.0
MarkupSafe==2.0.0
pipenv==2020.11.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
rope==0.18.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
virtualenv==20.4.3
virtualenv-clone==0.5.4
